v1.1.1 Refactor
Moving interface packages to separate Bukkit, Bungee and Common further.
