+++ Panther/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>PantherParent</artifactId>
         <groupId>com.ayanix</groupId>
-        <version>1.1.0-SNAPSHOT</version>
+        <version>1.1.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -15,25 +15,25 @@
         <dependency>
             <groupId>com.ayanix.panther</groupId>
             <artifactId>API</artifactId>
-            <version>1.1.0-SNAPSHOT</version>
+            <version>1.1.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>com.ayanix.panther</groupId>
             <artifactId>Common</artifactId>
-            <version>1.1.0-SNAPSHOT</version>
+            <version>1.1.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>com.ayanix.panther</groupId>
             <artifactId>Bukkit</artifactId>
-            <version>1.1.0-SNAPSHOT</version>
+            <version>1.1.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>com.ayanix.panther</groupId>
             <artifactId>Bungee</artifactId>
-            <version>1.1.0-SNAPSHOT</version>
+            <version>1.1.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
