+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.ayanix</groupId>
     <artifactId>PantherParent</artifactId>
-    <version>1.1.0-SNAPSHOT</version>
+    <version>1.1.1-SNAPSHOT</version>
     <name>PantherParent</name>
 
     <organization>
